feat(upload): implement image upload service with Cloudinary
- Add new upload controller and service to handle image uploads
- Move Cloudinary configuration and upload logic into dedicated service
- Update upload route to use new controller
- Remove outdated comment in App.jsx
